feat: update required Flowcore CLI version to 2.4.0 or higher
chore: update dataCore configuration for conference-app-dev

feat: add scenario configuration for conference-app

The commit updates the required Flowcore CLI version to 2.4.0 or higher. It also updates the dataCore configuration for conference-app-dev by adding a scenario configuration for conference-app. This scenario contains transformers for the open source conference app and includes events and parameters for the tickets transformer.